Added local files example
This adds and example for how to use colly to visit local files.
In order for this to work, a slight change was needed to the
http_backend to prevent panics when an http.Response does not
include the original request.

To run this example, you must run the acutal binary and not use
'go run' as the relative paths to the files will be different.