Adds HTTP Tracing to colly requests that are accessible from the colly Response.
Go HTTP supports tracing with the "net/http/httptrace" package. This adds a very
basic HTTPTrace object to the HTTP Request that tracks HTTP request timing. This
is useful for users wanting to tune their crawling based on RTT.

The resulting trace object is added to the colly Response which is accessible
to a caller via any of the Response callbacks.